feat: CurrentPrice volume, value Long에서 BigDecimal로 변경

+++ src/main/java/depth/finvibe/investment/modules/market/domain/CurrentPrice.java
@@ -5,6 +5,6 @@ import java.time.LocalDateTime;
 import java.util.Objects;
 
 public record CurrentPrice(Long stockId, LocalDateTime at, BigDecimal price, BigDecimal open, BigDecimal high, BigDecimal low, BigDecimal close,
-                           BigDecimal prevDayChangePct, Long volume, Long value) {
+                           BigDecimal prevDayChangePct, BigDecimal volume, BigDecimal value) {
 
 }
